+++ lib/ui/widgets/posts_list_item/list_item_content/export.dart
@@ -1,2 +1,3 @@
 export 'post_content.dart';
-export 'post_content_header.dart';
+export 'post_item_header.dart';
+export 'post_images_previewer.dart';