+++ lib/ui/widgets/posts_list_item/list_item_content/post_content.dart
@@ -2,8 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:mooncake/entities/entities.dart';
 import 'package:mooncake/ui/ui.dart';
 
-import 'post_content_header.dart';
-import 'post_content_images.dart';
+import 'post_item_header.dart';
+import 'post_images_previewer.dart';
 import 'post_content_message.dart';
 
 /// Contains the main content of a post. Such content is made of
@@ -12,11 +12,7 @@ import 'post_content_message.dart';
 /// - The image(s) associated to the post
 class PostContent extends StatelessWidget {
   final Post post;
-
-  const PostContent({
-    Key key,
-    @required this.post,
-  }) : super(key: key);
+  const PostContent({Key key, @required this.post}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -29,11 +25,10 @@ class PostContent extends StatelessWidget {
           PostMessage(key: PostsKeys.postItemMessage(post.id), post: post),
         if (post.message != null && post.message.isNotEmpty)
           const SizedBox(height: PostsTheme.defaultPadding),
-        if (post.images.isNotEmpty)
-          PostImagesPreviewer(
-            key: PostsKeys.postItemImagePreviewer(post.id),
-            post: post,
-          ),
+        PostImagesPreviewer(
+          key: PostsKeys.postItemImagePreviewer(post.id),
+          post: post,
+        ),
       ],
     );
   }